Create job to score forecasts

+++ db/migrate/20170618023158_create_forecasts.rb
@@ -3,7 +3,7 @@ class CreateForecasts < ActiveRecord::Migration[5.1]
     create_table :forecasts do |t|
       t.integer :high_temp
       t.integer :low_temp
-      t.float :precipitation
+      t.decimal :precipitation
       t.integer :wind
       t.float :score
       t.belongs_to :event, foreign_key: true
